Adding ability to get current URL from IWebBrowser2 if IHTMLDocument2 fails.
Under normal circumstances, the IE driver gets the current URL from the
document loaded in the browser. When IE browses to a non-HTML page, such
as a PDF file with the Adobe Reader plugin installed, getting the document
fails, since there is no "document" in the browser. In that case, fall
back to getting the LocationURL property of the IWebBrowser2 interface.
It's arguable that this is the wrong thing to do, since the driver is
useless without an actual document, but we'll let this work anyway. Fixes
issue #5361.